Revert the gcc sign conversion warning fixes.
This reverts commits 0cc03003281eb0d4c6924ef2ba80f52c852bb8ff to
38d15f8f90c8077bfd1bc27728cf51f18d97e327.

The 0cc03003 commit breaks MBROLA voice support. As such, there
may be other breakages in those commits.

+++ src/libespeak-ng/speech.c
@@ -237,7 +237,7 @@ ESPEAK_NG_API espeak_ng_STATUS espeak_ng_InitializeOutput(espeak_ng_OUTPUT_MODE 
 	return ENS_OK;
 }
 
-size_t GetFileLength(const char *filename)
+int GetFileLength(const char *filename)
 {
 	struct stat statbuf;
