+++ src/libespeak-ng/speech.h
@@ -75,7 +75,7 @@ int LookupMnem(MNEM_TAB *table, const char *string);
 extern char path_home[N_PATH_HOME];    // this is the espeak-data directory
 
 extern ESPEAK_NG_API void strncpy0(char *to, const char *from, int size);
-extern ESPEAK_NG_API size_t GetFileLength(const char *filename);
+extern ESPEAK_NG_API int  GetFileLength(const char *filename);
 char *Alloc(int size);
 void Free(void *ptr);
