+++ src/libespeak-ng/wave.h
@@ -55,7 +55,7 @@ extern void wave_set_callback_is_output_enabled(t_wave_callback *cb);
 
 // general functions
 extern void clock_gettime2(struct timespec *ts);
-extern void add_time_in_ms(struct timespec *ts, uint64_t time_in_ms);
+extern void add_time_in_ms(struct timespec *ts, int time_in_ms);
 
 // for tests
 extern void *wave_test_get_write_buffer();
